+++ projects/net.splitcells.website.server/src/main/java/net/splitcells/website/server/project/renderer/RendererMerger.java
@@ -15,6 +15,7 @@ import net.splitcells.dem.data.set.list.List;
 import net.splitcells.dem.data.set.list.Lists;
 import net.splitcells.dem.environment.config.StaticFlags;
 import net.splitcells.dem.lang.perspective.Perspective;
+import net.splitcells.website.server.config.Context;
 import net.splitcells.website.server.project.ProjectRenderer;
 import net.splitcells.website.server.project.RenderingResult;
 
@@ -46,14 +47,14 @@ public class RendererMerger implements Renderer {
     }
 
     @Override
-    public Optional<RenderingResult> renderFile(String path, ProjectRenderer projectRenderer) {
+    public Optional<RenderingResult> renderFile(String path, ProjectRenderer projectRenderer, Context context) {
         final var rendering = renderers.stream()
-                .map(e -> e.renderFile(path, projectRenderer))
+                .map(e -> e.renderFile(path, projectRenderer, context))
                 .filter(e -> e.isPresent())
                 .collect(Lists.toList());
         if (rendering.size() > 1) {
             final var matchedExtensions = renderers.stream()
-                    .filter(r -> r.renderFile(path, projectRenderer).isPresent())
+                    .filter(r -> r.renderFile(path, projectRenderer, context).isPresent())
                     .map(Object::toString)
                     .reduce((a, b) -> a + ", " + b)
                     .get();