+++ projects/net.splitcells.website.server/src/main/java/net/splitcells/website/server/project/renderer/XmlRenderer.java
@@ -5,6 +5,7 @@ import net.splitcells.dem.data.set.Sets;
 import net.splitcells.dem.lang.Xml;
 import net.splitcells.dem.lang.namespace.NameSpaces;
 import net.splitcells.dem.resource.Files;
+import net.splitcells.website.server.config.Context;
 import net.splitcells.website.server.project.FileStructureTransformer;
 import net.splitcells.website.server.project.ProjectRenderer;
 import net.splitcells.website.server.project.RenderingResult;
@@ -38,7 +39,7 @@ public class XmlRenderer implements Renderer {
     }
 
     @Override
-    public Optional<RenderingResult> renderFile(String path, ProjectRenderer projectRenderer) {
+    public Optional<RenderingResult> renderFile(String path, ProjectRenderer projectRenderer, Context context) {
         if (path.endsWith(".html")) {
             final var xmlFile = projectRenderer
                     .projectFolder()