+++ projects/net.splitcells.website.server/src/main/java/net/splitcells/website/server/project/renderer/commonmark/CommonMarkReadmeExtension.java
@@ -13,6 +13,7 @@ package net.splitcells.website.server.project.renderer.commonmark;
 import net.splitcells.dem.data.set.Set;
 import net.splitcells.dem.lang.perspective.Perspective;
 import net.splitcells.dem.resource.Files;
+import net.splitcells.website.server.config.Context;
 import net.splitcells.website.server.project.LayoutUtils;
 import net.splitcells.website.server.project.ProjectRenderer;
 import net.splitcells.website.server.project.renderer.Renderer;
@@ -43,13 +44,13 @@ public class CommonMarkReadmeExtension implements Renderer {
     }
 
     @Override
-    public Optional<RenderingResult> renderFile(String path, ProjectRenderer projectRenderer) {
+    public Optional<RenderingResult> renderFile(String path, ProjectRenderer projectRenderer, Context context) {
         final var readmePath = projectRenderer.projectFolder().resolve("README.md");
         if (path.endsWith("README.html")) {
             if (Files.is_file(readmePath)) {
                 final var pathContent = readString(readmePath);
                 return Optional.of(
-                        renderingResult(renderer.render(pathContent, projectRenderer, path)
+                        renderingResult(renderer.render(pathContent, projectRenderer, path, context)
                                 , TEXT_HTML.toString()));
             }
         }