+++ CHANGELOG.md
@@ -20,7 +20,12 @@ Links to such platforms may or may not be valid.
 Releases are done every time an important ticket is completed.
 ## [Unreleased]
 ### Major Changes
-* **2022-02-22** **\#s72** Remove unused class `net.splitcells.website.server.project.Location`.
+* **2022-02-22** **\#s72**:
+  1. Remove unused class `net.splitcells.website.server.project.Location`.
+  2. Define rendering context class, in order to pass additional or
+     configuration info to renderers.
+  3. Add context parameter to `ProjectRenderer#renderHtmlBodyContent`
+     and `Renderer#renderFile`.
 * **2022-02-21** **\#142** Remove `net.splitcells.website.html.content`
   project.
   It was an alternative minimal style for the website.